submissions/MindPulse/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML Libraries
anthropic==0.7.1
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2

# Data Processing
pandas==2.1.3
datasets==2.15.0

# Vector Store (for RAG)
chromadb==0.4.18
faiss-cpu==1.7.4

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging
loguru==0.7.2

